🚧 Rework text styles
