+++ src/App.tsx
@@ -2,7 +2,7 @@ import React, { Component, Fragment } from "react";
 import { Route, Switch } from "react-router-dom";
 import { ThemeProvider, createGlobalStyle } from "styled-components";
 
-import { ThemeColors, ThemeState } from "./reducers/theme";
+import { ThemeState } from "./reducers/theme";
 import ScrollToTop from "./components/ScrollToTop";
 import Header from "./components/Header";
 import AppSidebar from "./containers/AppSidebar";
@@ -17,6 +17,7 @@ import SubredditSearch from "./pages/SubredditSearch";
 import SavedPosts from "./pages/SavedPosts";
 import Popovers from "./components/Popovers";
 import UserPage from "./pages/UserPage";
+import { ThemeColors } from "./components/Base/theme";
 
 type GlobalStyleProps = {
   theme: ThemeColors;
